cmactions: Add torch action for ChopChop
This commit is based off the commit fb8b9382 by @crpalmer
  victara: cmactions: Make double-twist/chop-chop actions configurable
  Change-Id: I690427f29584592860b60f0b19321752c28bd104

This commit has been ammended to omit the IR silencer gesture, since
osprey and lux lack the IR sensors. Also, the ChopChop gesture has
been assigned specifically to the torch, and the camera twist gesture
has been assigned specifically to the camera. The extra configurability
is not necessary.