osprey: remove perfd [1/2]
It doesn't seem very useful, and it spams logcat with errors.

It provides a clumsy mechanism for certain audio/video routines to
temporarily request an elevated CPU frequency, and it also seems to be
used to ramp up frequencies on every touch, and keep retrying in a
loop if it fails. The perfd daemon creates the /dev/socket/mpctl socket
that the libqti-perfd-client.so library connects to. The library
provides the perf_lock_acq and perf_lock_rel functions to send requests
to the perfd daemon over the /dev/socket/mpctl socket. The perd daemon
listens for these requests, and temporarily constricts the governor
frequency scaling policy to an elevated frequency. This functionality
is used by various AV bits, as well as in
device/qcom/common/power/utils.c

In utils.c, it appears to try to repeatedly make these requests to
acquire elevated frequencies whenver you perform certain mundane
interactions (like pulling down the status bar drawer). This
functionality is useless, as CONFIG_CPU_BOOST in the kernel already
does this in a much more efficient manner. The PowerHAL tries to
call the function repeatedly and it just wastes battery with no
benefit to user experience.

To summarize, perfd is useless, as it's useful functionality is already
done better by CPU_BOOST in the kernel, and everything else is just
wasteful, clumsy, and unnecessary. We don't need this junk wasting
battery life.