Update patch set 1
Patch Set 1: Code-Review+1

I kept it after the screen density setting because that's how it was on the common repo and also because you choose to inherit a makefile or another based on density.
Anyway is a cosmetic change that won't matter. Feel free to make a Patch Set 2 for the three devices if it's that much important for you.

Patch-set: 1
Reviewer: Oliver Garcia Albertos <4733@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=+1